refactor(libdatadog): create namespaces for data-pipeline and logging
