+++ tracer/src/Datadog.Trace/Configuration/GlobalSettings.cs
@@ -9,6 +9,7 @@ using System;
 using Datadog.Trace.Configuration.ConfigurationSources.Telemetry;
 using Datadog.Trace.Configuration.Telemetry;
 using Datadog.Trace.LibDatadog;
+using Datadog.Trace.LibDatadog.Logging;
 using Datadog.Trace.Logging;
 using Datadog.Trace.SourceGenerators;
 using Datadog.Trace.Telemetry;
@@ -108,7 +109,7 @@ namespace Datadog.Trace.Configuration
                 DatadogLogging.UseDefaultLevel();
             }
 
-            LibDatadog.Logger.Instance.SetLogLevel(debugEnabled: enabled);
+            Logger.Instance.SetLogLevel(debugEnabled: enabled);
 
             TelemetryFactory.Config.Record(ConfigurationKeys.DebugEnabled, enabled, ConfigurationOrigins.Code);
         }
@@ -125,7 +126,7 @@ namespace Datadog.Trace.Configuration
             var isLibdatadogAvailable = LibDatadogAvailaibilityHelper.IsLibDatadogAvailable;
             if (isLibdatadogAvailable.IsAvailable)
             {
-                LibDatadog.Logger.Instance.SetLogLevel(debugEnabled: false);
+                Logger.Instance.SetLogLevel(debugEnabled: false);
             }
 
             GlobalConfigurationSource.Reload(isLibdatadogAvailable.IsAvailable);
