+++ tracer/src/Datadog.Trace/LibDatadog/DataPipeline/TraceExporter.cs
@@ -7,8 +7,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Globalization;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 using Datadog.Trace.Agent;
@@ -16,7 +14,7 @@ using Datadog.Trace.Logging;
 using Datadog.Trace.Vendors.Newtonsoft.Json;
 using static Datadog.Trace.Agent.Api;
 
-namespace Datadog.Trace.LibDatadog;
+namespace Datadog.Trace.LibDatadog.DataPipeline;
 
 internal class TraceExporter : SafeHandle, IApi
 {
