+++ tracer/src/Datadog.Trace/LibDatadog/DataPipeline/TraceExporterError.cs
@@ -8,7 +8,7 @@
 using System;
 using System.Runtime.InteropServices;
 
-namespace Datadog.Trace.LibDatadog;
+namespace Datadog.Trace.LibDatadog.DataPipeline;
 
 /// <summary>
 /// Represents errors that can occur when exporting traces.
