+++ tracer/src/Datadog.Trace/LibDatadog/DataPipeline/TraceExporterException.cs
@@ -8,7 +8,7 @@
 using System;
 using System.Runtime.InteropServices;
 
-namespace Datadog.Trace.LibDatadog;
+namespace Datadog.Trace.LibDatadog.DataPipeline;
 
 /// <summary>
 /// Represents an exception thrown by the libdatadog library.
