+++ tracer/src/Datadog.Trace/LibDatadog/DataPipeline/TraceExporterInputFormat.cs
@@ -5,7 +5,7 @@
 
 #nullable enable
 
-namespace Datadog.Trace.LibDatadog;
+namespace Datadog.Trace.LibDatadog.DataPipeline;
 
 /// <summary>
 /// Represents the format of the input traces, as expected by the trace exporter.
